requirements.txt
lm-eval>=0.4.0
tabulate>=0.9.0
openai>=1.0.0
goodfire
python-dotenv>=1.0.0
lm-eval[api]
